archive/old_malibu/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "malibu"
version = "0.1.0"
description = "Malibu — AI-powered terminal coding agent with rich TUI, Git integration, hooks, and plugin system."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["agent", "terminal", "coding-agent", "tui", "ai-agents", "langgraph", "acp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # ACP Protocol
    "agent-client-protocol>=0.8.0",
    # LangChain 1.0 LTS + LangGraph 1.0
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langgraph>=1.0,<2.0",
    "langsmith>=0.3.0",
    "deepagents",
    # Model providers (install those you need)
    "langchain-openai",
    "langchain-anthropic",
    # Native tool runtime
    "openai>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "playwright>=1.40.0",
    "tenacity>=8.2.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "ddgs>=9.10.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pillow>=10.0.0",
    "pymupdf>=1.26.7",
    "multilspy>=0.0.15",
    "numpy>=1.24.0",
    "tavily-python>=0.5.0",
    "google-genai>=0.1.0",
    "google-cloud-storage>=2.18.2",
    "google-cloud-aiplatform>=1.70.0",
    # Database (PostgreSQL + async)
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    # Authentication
    "pyjwt[crypto]>=2.9.0",
    "bcrypt>=4.2.0",
    # Configuration
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # Observability
    "structlog>=24.4.0",
    # Terminal UI
    "textual>=1.0",
]

[project.optional-dependencies]
logfire = ["logfire>=0.14"]

[project.scripts]
malibu = "malibu.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.0",
    "ruff>=0.9.0",
    "ty>=0.0.1",
    "aiosqlite>=0.20.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "S", "A", "UP", "YTT"]
ignore = ["S101", "S603", "S607"]

[tool.ruff.lint.isort]
known-first-party = ["malibu"]

[tool.ty.environment]
python-version = "3.11"
